+++ webClient/src/app/shared/editor-control/editor-control.service.ts
@@ -1272,6 +1272,10 @@ export class EditorControlService implements ZLUX.IEditor, ZLUX.IEditorMultiBuff
         monaco.editor.setTheme('ceedump-dark');
         break;
       }
+      case 'hlasm': {
+        monaco.editor.setTheme('hlasm-dark');
+        break;
+      }
       default: {
         if (this._defaultTheme) {
           monaco.editor.setTheme(this._defaultTheme);