+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `3.5.0`
 
 - Enhancement: Added CEEDUMP language mode. Files named `CEEDUMP.*` are automatically detected. It provides syntax highlighting and hover-help for headers, common messages, registers and more. 32-bit and 64-bit hex addresses are colored by byte significance to aid in pointer analysis.
+- Enhancement: HLASM language mode now has hover-help and improved syntax highlighting
   
 ## `3.4.0`